+++ src/conf/mod.rs
@@ -1,11 +1,23 @@
 pub mod conf_tree;
+
 pub mod fn_conf_keywd;
+
 pub mod fn_config;
+
 pub mod fn_conf_kind;
+
 pub mod metric_config;
+
 pub mod task_config;
+
 pub mod conf_duration;
+
 pub mod api_client_config;
+
 pub mod conf_keywd;
+
 pub mod tcp_client_config;
+
 pub mod multi_queue_config;
+
+pub mod service_config;
